pyproject.toml
[project]
name = "agent-for-interviews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Базовые зависимости
    "pydantic==2.10.6",
    "pydantic_core==2.27.2",
    "typing_extensions==4.12.2",
    "python-dotenv==1.0.1",
    "rich==13.9.4",
    "tqdm==4.67.1",
    "requests==2.32.3",
    "urllib3==2.3.0",
    "PyYAML==6.0.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "idna==3.10",
    "anyio==4.8.0",
    "sniffio==1.3.1",
    "tiktoken>=0.9.0",
]

[project.optional-dependencies]
# Зависимости для работы с AI и чат-ботами
ai = [
    "gigachat==0.1.39",
    "openai==1.65.5",
    "langchain-core==0.3.44",
    "langchain-gigachat==0.3.6",
    "langsmith==0.3.13",
    "tenacity==9.0.0",
]

# Зависимости для распознавания и записи звука
audio = [
    "PyAudio==0.2.14",
    "SpeechRecognition==3.14.1",
]

# Зависимости для работы со скриншотами и изображениями
screenshot = [
    "pillow==11.1.0",
    "PyAutoGUI==0.9.54",
    "PyGetWindow==0.0.9",
    "PyScreeze==1.0.1",
    "MouseInfo==0.1.3",
    "PyMsgBox==1.0.9",
    "PyRect==0.2.0",
    "pytesseract==0.3.13",
    "pyperclip==1.9.0",
    "pytweening==1.2.0",
]

# Зависимости для macOS (специфичные для платформы)
macos = [
    "pyobjc-core==11.0",
    "pyobjc-framework-Cocoa==11.0",
    "pyobjc-framework-Quartz==11.0",
    "rubicon-objc==0.5.0",
]

# Зависимости для HTTP клиентов
http = [
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "h11==0.14.0",
    "requests-toolbelt==1.0.0",
]

# Зависимости для форматирования и документации
format = [
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "Pygments==2.19.1",
]

# Зависимости для разработки
dev = [
    "ruff>=0.11.0",
    "types-requests==2.32.0.20250306",
]

# Все зависимости вместе
all = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "distro==1.9.0",
    "gigachat==0.1.39",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.9.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-core==0.3.44",
    "langchain-gigachat==0.3.6",
    "langsmith==0.3.13",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "MouseInfo==0.1.3",
    "openai==1.65.5",
    "orjson==3.10.15",
    "packaging==24.2",
    "pillow==11.1.0",
    "PyAudio==0.2.14",
    "PyAutoGUI==0.9.54",
    "pydantic==2.10.6",
    "pydantic_core==2.27.2",
    "PyGetWindow==0.0.9",
    "Pygments==2.19.1",
    "PyMsgBox==1.0.9",
    "pyobjc-core==11.0",
    "pyobjc-framework-Cocoa==11.0",
    "pyobjc-framework-Quartz==11.0",
    "pyperclip==1.9.0",
    "PyRect==0.2.0",
    "PyScreeze==1.0.1",
    "pytesseract==0.3.13",
    "python-dotenv==1.0.1",
    "pytweening==1.2.0",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rubicon-objc==0.5.0",
    "ruff>=0.11.0",
    "sniffio==1.3.1",
    "SpeechRecognition==3.14.1",
    "tenacity==9.0.0",
    "tqdm==4.67.1",
    "types-requests==2.32.0.20250306",
    "typing_extensions==4.12.2",
    "urllib3==2.3.0",
    "zstandard==0.23.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
target-version = "py312"
